Create `subspace-farmer-components` crate where plotting, auditing and proving code is moved for wider reuse
