+++ crates/subspace-farmer-components/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+name = "subspace-farmer-components"
+description = "Farmer for the Subspace Network Blockchain"
+license = "MIT OR Apache-2.0"
+version = "0.1.0"
+authors = ["Nazar Mokrynskyi <nazar@mokrynskyi.com>"]
+edition = "2021"
+include = [
+    "/src",
+    "/Cargo.toml",
+    "/README.md",
+]
+
+[lib]
+# Necessary for CLI options to work on benches
+bench = false
+
+[dependencies]
+async-trait = "0.1.57"
+bitvec = "1.0.1"
+blake2-rfc = "0.2.18"
+fs2 = "0.4.3"
+libc = "0.2.131"
+parity-scale-codec = "3.1.5"
+rand = "0.8.5"
+schnorrkel = "0.9.1"
+static_assertions = "1.1.0"
+subspace-solving = { version = "0.1.0", path = "../subspace-solving" }
+subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives" }
+subspace-rpc-primitives = { version = "0.1.0", path = "../subspace-rpc-primitives" }
+subspace-verification = { version = "0.1.0", path = "../subspace-verification" }
+thiserror = "1.0.32"
+tracing = "0.1.36"
+
+[dev-dependencies]
+criterion = "0.4.0"
+futures = "0.3.21"
+memmap2 = "0.5.7"
+rayon = "1.5.3"
+subspace-archiving = { version = "0.1.0", path = "../subspace-archiving" }
+
+[[bench]]
+name = "plotting"
+harness = false
+
+[[bench]]
+name = "auditing"
+harness = false
+
+[[bench]]
+name = "proving"
+harness = false
