+++ crates/subspace-farmer-components/README.md
@@ -1,0 +1,5 @@
+# Subspace Farmer Components
+
+Components of the reference implementation of Subspace Farmer for Subspace Network Blockchain.
+
+These components are used to implement farmer itself, but can also be used independently if necessary.
