+++ crates/subspace-farmer/src/lib.rs
@@ -30,8 +30,6 @@
 //! are `target ± ½ * solution range` (while also handing overflow/underflow) when interpreted as
 //! 64-bit unsigned integers.
 
-#[doc(hidden)]
-pub mod file_ext;
 pub(crate) mod identity;
 pub(crate) mod object_mappings;
 pub mod reward_signing;
